requirements.txt
streamlit==1.29.0
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
Faker==21.0.0
python-dotenv==1.0.0
openpyxl==3.1.2